moved quarterly_usage to commands
