+++ dds_web/__init__.py
@@ -269,6 +269,7 @@ def create_app(testing=False, database_uri=None):
             set_available_to_expired,
             set_expired_to_archived,
             delete_invites,
+            quarterly_usage,
         )
 
         app.cli.add_command(fill_db_wrapper)
@@ -280,6 +281,7 @@ def create_app(testing=False, database_uri=None):
         app.cli.add_command(set_available_to_expired)
         app.cli.add_command(set_expired_to_archived)
         app.cli.add_command(delete_invites)
+        app.cli.add_command(quarterly_usage)
 
         # Make version available inside jinja templates:
         @app.template_filter("dds_version")
